requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
scipy>=1.7.0
coco-experiment
cocopp
cma>=3.3.0
scikit-optimize>=0.9.0